Rename rule-no-duplicate-properties
Fix rebase

Fix rebase2
