+++ docs/user-guide/example-config.md
@@ -31,6 +31,7 @@ You might want to learn a little about [the conventions](https://github.com/styl
     "custom-property-pattern": string,
     "declaration-bang-space-after": "always"|"never",
     "declaration-bang-space-before": "always"|"never",
+    "declaration-block-no-duplicate-properties": true,
     "declaration-block-no-shorthand-property-overrides": true,
     "declaration-block-no-single-line": true,
     "declaration-block-semicolon-newline-after": "always"|"always-multi-line"|"never-multi-line",
@@ -89,7 +90,6 @@ You might want to learn a little about [the conventions](https://github.com/styl
     "property-whitelist": string|[],
     "root-no-standard-properties": true,
     "rule-nested-empty-line-before": "always"|"never",
-    "rule-no-duplicate-properties": true,
     "rule-non-nested-empty-line-before": "always"|"never",
     "rule-properties-order": "alphabetical"|[],
     "selector-class-pattern": string,
