+++ src/rules/index.js
@@ -23,6 +23,7 @@ import customPropertyNoOutsideRoot from "./custom-property-no-outside-root"
 import customPropertyPattern from "./custom-property-pattern"
 import declarationBangSpaceAfter from "./declaration-bang-space-after"
 import declarationBangSpaceBefore from "./declaration-bang-space-before"
+import declarationBlockNoDuplicateProperties from "./declaration-block-no-duplicate-properties"
 import declarationBlockNoShorthandPropertyOverrides from "./declaration-block-no-shorthand-property-overrides"
 import declarationBlockNoSingleLine from "./declaration-block-no-single-line"
 import declarationBlockSemicolonNewlineAfter from "./declaration-block-semicolon-newline-after"
@@ -142,6 +143,7 @@ export default {
   "declaration-bang-space-after": declarationBangSpaceAfter,
   "declaration-bang-space-before": declarationBangSpaceBefore,
   "declaration-block-no-shorthand-property-overrides": declarationBlockNoShorthandPropertyOverrides,
+  "declaration-block-no-duplicate-properties": declarationBlockNoDuplicateProperties,
   "declaration-block-no-single-line": declarationBlockNoSingleLine,
   "declaration-block-semicolon-newline-after": declarationBlockSemicolonNewlineAfter,
   "declaration-block-semicolon-newline-before": declarationBlockSemicolonNewlineBefore,
